chore(mv-sign-in): move sign in logic into SIWE client
